fixed bug with ArcticLibraryBinding not picking up the arctic reset, improved test coverage, split the multithreading tests to separate files
